backend-python/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-cloud-aiplatform==1.30.0
neo4j==5.8.0
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.7